Add and update SettingsTests
[ignore-commit-lint]
