+++ Parent/Parent/Profile/Settings/View/ProfileSettingsView.swift
@@ -38,6 +38,7 @@ public struct ProfileSettingsView: View {
         }
         .background(Color.backgroundLightest)
         .navigationBarTitleView(String(localized: "Settings", bundle: .core))
+        .navigationBarStyle(.modal)
     }
 }
